+++ app/(admin)/layout.jsx
@@ -2,7 +2,6 @@ import { getServerSession } from "next-auth";
 import { redirect } from "next/navigation";
 import { serverAuthOptions } from "@/lib/auth";
 import AdminWrapper from "@/wrappers/AdminWrapper";
-import { AdminProvider } from "@/context/AdminContext";
 
 export default async function ProtectedLayout({ children }) {
   const session = await getServerSession(serverAuthOptions);