+++ wrappers/AdminWrapper.js
@@ -1,7 +1,0 @@
-"use client";
-
-import { AdminProvider } from "@/context/AdminContext";
-
-export default function AdminWrapper({ children, isAdmin }) {
-  return <AdminProvider isAdmin={isAdmin}>{children}</AdminProvider>;
-}